+++ include/exec/helper-tcg.h
@@ -62,7 +62,6 @@
 #include "helper.h"
 #include "trace/generated-helpers.h"
 #include "tcg-runtime.h"
-#include "tcg-runtime-sym.h"
 #include "tcg-runtime-symsan.h"
 
 #undef str